requirements.txt
# BOQTenders - Requirements
# Python 3.9+

# ============================================
# Core Dependencies
# ============================================
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
PyYAML>=6.0.0

# ============================================
# LLM & AI
# ============================================
# LangChain 0.1.x (Classic API - stable)
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.1,<0.1.0
langchain-google-genai>=0.0.5

# Vector Store
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.0

# ============================================
# PDF Processing
# ============================================
PyPDF2>=3.0.0
requests>=2.31.0

# ============================================
# Web Framework
# ============================================
# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# UI
streamlit>=1.28.0

# ============================================
# Development (optional)
# ============================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.5.0
